feature: Reworked the entertainment poster card component and renamed some of the existing files
