+++ src/components/Cards/Entertainment/Poster/Poster.stories.js
@@ -1,0 +1,135 @@
+import * as React from 'react';
+import { action } from '@storybook/addon-actions';
+
+import PosterCard from './Poster';
+import Preview from '../../../Blocks/Storybook/Preview';
+import { GroupedExample, SingleExample } from '../../Shared';
+
+export const Default = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard />
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const Title = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard title='The Walking Dead' />
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const ReleaseDate = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard releaseDate='2010' />
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const Rating = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard rating={75} />
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const image = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard image='https://image.tmdb.org/t/p/original/lAwLmgq1zy0xJnusszmvWTLjFlO.jpg' />
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const onClick = () => (
+	<Preview
+		content={
+			<div className='flex'>
+				<PosterCard onClick={action('Poster card has been clicked')} />{' '}
+			</div>
+		}
+		background='#F5F5F5'
+	/>
+);
+
+export const GroupedPosters = () => (
+	<div>
+		<Preview
+			content={
+				<>
+					<div>
+						<h1>Popular Movies</h1>
+					</div>
+					<div className='flex flex-wrap'>
+						<PosterCard
+							title='Westworld'
+							image='https://image.tmdb.org/t/p/original/y55oBgf6bVMI7sFNXwJDrSIxPQt.jpg'
+							rating={75}
+						/>
+						<PosterCard
+							title='Westworld'
+							image='https://image.tmdb.org/t/p/original/y55oBgf6bVMI7sFNXwJDrSIxPQt.jpg'
+							rating={75}
+						/>
+						<PosterCard
+							title='Westworld'
+							image='https://image.tmdb.org/t/p/original/y55oBgf6bVMI7sFNXwJDrSIxPQt.jpg'
+							rating={75}
+						/>
+						<PosterCard
+							title='Westworld'
+							image='https://image.tmdb.org/t/p/original/y55oBgf6bVMI7sFNXwJDrSIxPQt.jpg'
+							rating={75}
+						/>
+
+						<PosterCard
+							title='The Walking Dead'
+							image='https://image.tmdb.org/t/p/original/5l10EjdgPxu8Gbl5Ww6SWkVQH6T.jpg'
+							rating={25}
+						/>
+						<PosterCard
+							title='The Walking Dead'
+							image='https://image.tmdb.org/t/p/original/5l10EjdgPxu8Gbl5Ww6SWkVQH6T.jpg'
+							rating={25}
+						/>
+						<PosterCard
+							title='The Walking Dead'
+							image='https://image.tmdb.org/t/p/original/5l10EjdgPxu8Gbl5Ww6SWkVQH6T.jpg'
+							rating={25}
+						/>
+						<PosterCard
+							title='The Walking Dead'
+							image='https://image.tmdb.org/t/p/original/5l10EjdgPxu8Gbl5Ww6SWkVQH6T.jpg'
+							rating={25}
+						/>
+					</div>
+				</>
+			}
+			background='white'
+		/>
+	</div>
+);
+
+export default {
+	title: ' Design System/Cards/Entertainment/Poster',
+	component: PosterCard
+};
