+++ tailwind.config.js
@@ -7,6 +7,9 @@ module.exports = {
 				primary: 'var(--primary)',
 				secondary: 'var(--secondary)',
 				tertiary: 'var(--tertiary)'
+			},
+			boxShadow: {
+				card: '0 2px 8px rgba(0, 0, 0, 0.1)'
 			}
 		}
 	},
